+++ src/pages/_app.tsx
@@ -1,9 +1,8 @@
 import '@/styles/globals.css';
 import type { AppProps } from 'next/app';
 import { Commissioner } from 'next/font/google';
-import { Header } from './home/components/Header';
-import { Footer } from './home/components/footer';
-
+import { Header } from '../components/header';
+import { Footer } from '../components/footer';
 
 const comissioner = Commissioner({
   subsets: ['latin'],
